SwiftLint added

+++ VKMessenger/ViewControllers/FriendsViewController.swift
@@ -12,10 +12,10 @@ class FriendsViewController: UITableViewController {
     private let cellID = "friend"
     private let networkManager = NetworkManager.shared
     private let storageManager = StorageManager.shared
-    
+
     private var items: [FriendItem] = []
     private var allFriends: [Friend] = []
-    
+
     override func viewDidLoad() {
         super.viewDidLoad()
         setupRefreshControl()
@@ -25,13 +25,12 @@ class FriendsViewController: UITableViewController {
             target: self,
             action: #selector(profileButtonPressed)
         )
-        
         navigationItem.rightBarButtonItem?.tintColor = Interface.textColor
         tableView.register(FriendViewCell.self, forCellReuseIdentifier: cellID)
         fetchData()
         tableView.rowHeight = 130
     }
-    
+
     @objc private func profileButtonPressed() {
         let animation = CATransition()
         animation.timingFunction = CAMediaTimingFunction(name: .easeOut)
@@ -40,22 +39,25 @@ class FriendsViewController: UITableViewController {
         navigationController?.view.layer.add(animation, forKey: nil)
         navigationController?.pushViewController(ProfileViewController(), animated: true)
     }
-    
+
     private func fetchData() {
-        let friendURL = URL(string: "https://api.vk.com/method/friends.get?user_ids=\(NetworkManager.userID)&fields=bdays,city,photo_200_orig,online&access_token=\(NetworkManager.token)&v=5.131")!
-        
+        guard let friendURL = URL(
+            string: "https://api.vk.com/method/friends.get?user_ids=\(NetworkManager.userID)" +
+            "&fields=bdays,city,photo_200_orig,online&access_token=\(NetworkManager.token)&v=5.131"
+        ) else { return }
+
         networkManager.fetch(FriendResponse.self, from: friendURL) { [weak self] result in
             switch result {
             case .success(let response):
                 self?.items = response.response.items
                 self?.fetchFriends()
-                
+
                 self?.allFriends.forEach { friend in
                     self?.storageManager.deleteFriend(friend)
                 }
-                
+
                 self?.allFriends = []
-                
+
                 self?.items.forEach { item in
                     self?.storageManager.createFriend(
                         id: item.id,
@@ -66,9 +68,9 @@ class FriendsViewController: UITableViewController {
                         photoTwoHundred: item.photoTwoHundred ?? ""
                     )
                 }
-                
+
                 self?.fetchFriends()
-                
+
                 DispatchQueue.main.async {
                     self?.tableView.reloadData()
                     if self?.refreshControl != nil {
@@ -80,8 +82,8 @@ class FriendsViewController: UITableViewController {
             }
         }
     }
-    
-    private func fetchFriends() {
+
+    func fetchFriends() {
         storageManager.fetchFriends { result in
             switch result {
             case .success(let friends):
@@ -91,11 +93,11 @@ class FriendsViewController: UITableViewController {
             }
         }
     }
-    
+
     private func setupRefreshControl() {
         refreshControl = UIRefreshControl()
         refreshControl?.attributedTitle = NSAttributedString(string: "Pull to refresh")
-        
+
         let refreshAction = UIAction { [unowned self] _ in
             fetchData()
         }
@@ -105,28 +107,27 @@ class FriendsViewController: UITableViewController {
 
 // MARK: - UITableViewDataSource
 extension FriendsViewController {
-    
+
     override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
         allFriends.count
     }
-    
+
     override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
         let cell = tableView.dequeueReusableCell(withIdentifier: cellID, for: indexPath)
         guard let cell = cell as? FriendViewCell else { return UITableViewCell() }
-        
+
         cell.selectionStyle = .none
         cell.backgroundColor = Interface.cellColor
         let friend = allFriends[indexPath.row]
         cell.configure(with: friend)
-        
+
         return cell
     }
-    
+
     override func tableView(_ tableView: UITableView, didSelectRowAt indexPath: IndexPath) {
         let chatVC = ChatViewController()
         chatVC.id = items[indexPath.row].id
-        
+
         navigationController?.pushViewController(chatVC, animated: true)
     }
 }
-
