+++ VKMessenger/ViewControllers/GroupsViewController.swift
@@ -8,14 +8,14 @@
 import UIKit
 
 final class GroupsViewController: UITableViewController {
-    
+
     private let cellID = "group"
     private let networkManager = NetworkManager.shared
     private let storageManager = StorageManager.shared
-    
+
     private var groups: [GroupItem] = []
     private var allGroups: [Group] = []
-    
+
     override func viewDidLoad() {
         super.viewDidLoad()
         tableView.register(GroupViewCell.self, forCellReuseIdentifier: cellID)
@@ -26,34 +26,37 @@ final class GroupsViewController: UITableViewController {
     override func tableView(_ tableView: UITableView, numberOfRowsInSection section: Int) -> Int {
         groups.count
     }
-    
+
     override func tableView(_ tableView: UITableView, cellForRowAt indexPath: IndexPath) -> UITableViewCell {
         let cell = tableView.dequeueReusableCell(withIdentifier: cellID, for: indexPath)
         guard let cell = cell as? GroupViewCell else { return UITableViewCell() }
-        
+
         cell.selectionStyle = .none
         cell.backgroundColor = Interface.cellColor
         let group = allGroups[indexPath.row]
         cell.configure(with: group)
-        
+
         return cell
     }
-    
+
     private func fetchData() {
-        let groupUrl = URL(string: "https://api.vk.com/method/groups.get?user_id=\(NetworkManager.userID)&extended=1&fields=description&access_token=\(NetworkManager.token)&v=5.131")!
-        
+        guard let groupUrl = URL(
+            string: "https://api.vk.com/method/groups.get?user_id=\(NetworkManager.userID)" +
+            "&extended=1&fields=description&access_token=\(NetworkManager.token)&v=5.131"
+        ) else { return }
+
         networkManager.fetch(GroupResponse.self, from: groupUrl) { [weak self] result in
             switch result {
             case .success(let group):
                 self?.groups = group.response.items
                 self?.fetchGroups()
-                
+
                 self?.allGroups.forEach { group in
                     self?.storageManager.deleteGroup(group)
                 }
-                
+
                 self?.allGroups = []
-                
+
                 self?.groups.forEach { group in
                     self?.storageManager.createGroup(
                         id: group.id,
@@ -62,7 +65,7 @@ final class GroupsViewController: UITableViewController {
                         photoTwoHundred: group.photoOneHundred
                     )
                 }
-                
+
                 self?.fetchGroups()
                 DispatchQueue.main.async {
                     self?.tableView.reloadData()
@@ -72,7 +75,7 @@ final class GroupsViewController: UITableViewController {
             }
         }
     }
-    
+
     private func fetchGroups() {
         storageManager.fetchGroups { result in
             switch result {
